Implement Object#getClass()
To do so, we store the class object for a reference type inside
its dispatch vector.

This currently does not work well for array types, since the
class object associated with arrays is polyllvm.runtime.Array,
regardless of the base time and number of dimensions. To fix this,
we'll need to create a new dispatch vector for each array type.