Fix MissingDependencyExceptions
This fixes a thorny issue in which MissingDependencyExceptions were thrown
when trying to search for methods/fields within desugar transformations.
These exceptions were being silently caught by the scheduler. This is
expected behavior, but it leads to subtle bugs since some of the global
desugar transformations are not idempotent.

For now, the missing dependencies are added as explicit prerequisites
for the desugaring goal. Perhaps there could be a more automatic solution
in the future.