Branch from finally block using switch (no copies)
Previously, a finally block was copied as many times as necessary to
ensure that it could jump to the correct destination based
on how it was entered. This caused some code bloat, though was
a reasonable tradeoff.

However, duplicating a subtree of the AST can lead to other problems because,
even when making a deep copy, there are still aliased LocalInstance
references (for example). It's completely possible to fix these problems
by writing a custom deep copy pass, but, incidentally, rewriting the way
finally blocks were translated seemed easier.

So, we now store a "destination ID" before jumping to a finally block,
and we build a switch at the end of the finally block to jump to the
correct destination based on the stored destination ID.