Fix subtle generics bug regarding raw class types
Polyglot uses a CachingTransformingList to enumerate the methods
of a raw class type, but this data structures is not designed
to handle changes to its underlying list. Since desugaring
transformations can add methods to class types, this would
lead to IndexOutOfBounds exceptions within Polyglot when
accessing the methods of a raw class type.

The solution submitted here should be viewed as a workaround.