Workaround for file manager issue
The default file manager will try to load classes by reflection
if they are not found through the normal class path (apparently
as a workaround for issues related to JDK 9). This doesn't always work
for PolyLLVM, possibly related to the fact that we're running
PolyLLVM with JDK 8 yet using it to compile JDK 7.

The easy workaround for now is to override
ExtFileManager#getJavaFileForInput.