Implement registerNatives; dynamic native lookup
We now generate a stub for each native method. The stub
calls into the runtime to retrieve a function pointer for the
"real" native method implementation. The runtime finds this
pointer using either dlsym or by looking to see if it has
been registered using registerNatives(). Finally, the stub
calls the function pointer, forwarding all arguments.

Fixes #48, as a side effect.