Improve Linux support (#59)
Add appropriate clang/linker flags to get shared libraries working on Linux. Unit tests now work with `jdk-lite` on Linux.

Linking with the full JDK 7 still does not work on Linux due to undefined symbols, e.g.
"libjava.so: undefined reference to `JVM_GetArrayLength@SUNWprivate_1.1'".
It's unclear why that symbol reference exists on Linux, yet is not a problem on Mac OS X.
At some point we really might want to compile OpenJDK native code from source ourselves, so that we don't have to link with the system JDK in the first place.